+++ src/Controller/OrdreAchatController.php
@@ -20,6 +20,14 @@ class OrdreAchatController extends AbstractController
      */
     public function index(OrdreAchatRepository $ordreAchatRepository): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+        $userRoles = $this->getUser()->getRoles();
+
+//        Rejeter l'accès au non admin et non super admin
+        if (!(in_array('ROLE_SUPER_ADMIN', $userRoles) || in_array('ROLE_ADMIN', $userRoles))) {
+            $this->denyAccessUnlessGranted('ROLE_ADMIN');
+        }
+
         return $this->render('ordre_achat/index.html.twig', [
             'ordre_achats' => $ordreAchatRepository->findAll(),
         ]);
@@ -30,6 +38,14 @@ class OrdreAchatController extends AbstractController
      */
     public function new(Request $request): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+        $userRoles = $this->getUser()->getRoles();
+
+//        Rejeter l'accès non admin et non super admin
+        if (!(in_array('ROLE_SUPER_ADMIN', $userRoles) || in_array('ROLE_ADMIN', $userRoles))) {
+            $this->denyAccessUnlessGranted('ROLE_ADMIN');
+        }
+
         $ordreAchat = new OrdreAchat();
         $form = $this->createForm(OrdreAchatType::class, $ordreAchat);
         $form->handleRequest($request);
@@ -53,6 +69,14 @@ class OrdreAchatController extends AbstractController
      */
     public function show(OrdreAchat $ordreAchat): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+        $userRoles = $this->getUser()->getRoles();
+
+//        Rejeter l'accès au non admin et non super admin
+        if (!(in_array('ROLE_SUPER_ADMIN', $userRoles) || in_array('ROLE_ADMIN', $userRoles))) {
+            $this->denyAccessUnlessGranted('ROLE_ADMIN');
+        }
+
         return $this->render('ordre_achat/show.html.twig', [
             'ordre_achat' => $ordreAchat,
         ]);
@@ -63,6 +87,14 @@ class OrdreAchatController extends AbstractController
      */
     public function edit(Request $request, OrdreAchat $ordreAchat): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+        $userRoles = $this->getUser()->getRoles();
+
+//        Rejeter l'accès au non admin et non super admin
+        if (!(in_array('ROLE_SUPER_ADMIN', $userRoles) || in_array('ROLE_ADMIN', $userRoles))) {
+            $this->denyAccessUnlessGranted('ROLE_ADMIN');
+        }
+
         $form = $this->createForm(OrdreAchatType::class, $ordreAchat);
         $form->handleRequest($request);
 
@@ -83,6 +115,14 @@ class OrdreAchatController extends AbstractController
      */
     public function delete(Request $request, OrdreAchat $ordreAchat): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+        $userRoles = $this->getUser()->getRoles();
+
+//        Rejeter l'accès non admin et non super admin
+        if (!(in_array('ROLE_SUPER_ADMIN', $userRoles) || in_array('ROLE_ADMIN', $userRoles))) {
+            $this->denyAccessUnlessGranted('ROLE_ADMIN');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$ordreAchat->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($ordreAchat);
